requirements.txt
flask
werkzeug
keras
numpy
opencv-python
pillow
ultralytics
PyQt5
matplotlib
pandas
